Downstream of opm-common filesystem reorganisation
