+++ opm/material/fluidmatrixinteractions/EclHysteresisTwoPhaseLawParams.hpp
@@ -31,8 +31,8 @@
 #include "EclEpsScalingPoints.hpp"
 
 #if HAVE_ECL_INPUT
-#include <opm/parser/eclipse/Deck/Deck.hpp>
-#include <opm/parser/eclipse/EclipseState/EclipseState.hpp>
+#include <opm/input/eclipse/Deck/Deck.hpp>
+#include <opm/input/eclipse/EclipseState/EclipseState.hpp>
 #endif
 
 #include <string>
