+++ opm/material/fluidsystems/blackoilpvt/OilPvtMultiplexer.hpp
@@ -34,8 +34,8 @@
 #include "BrineCo2Pvt.hpp"
 
 #if HAVE_ECL_INPUT
-#include <opm/parser/eclipse/EclipseState/EclipseState.hpp>
-#include <opm/parser/eclipse/EclipseState/Runspec.hpp>
+#include <opm/input/eclipse/EclipseState/EclipseState.hpp>
+#include <opm/input/eclipse/EclipseState/Runspec.hpp>
 #endif
 
 namespace Opm {
